requirements.txt
streamlit>=1.31.0
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
python-dotenv>=1.0.0
pandas-ta>=0.3.14
